+++ Source/src/ui/PhiColours.h
@@ -13,10 +13,10 @@
 class PhiColourIds {
     // 256 colours for each class
     enum class ColourClass {
-        General = 0x000,
-        Module = 0x100,
-        Connection = 0x200,
-        Port = 0x300,
+        General = 0x000,     // IDs 0 - 255
+        Module = 0x100,      // IDs 256 - 511
+        Connection = 0x200,  // IDs 512 - 767
+        Port = 0x300,        // IDs 768 - 1023
     };
 public:
     struct General {enum Val{
@@ -50,102 +50,6 @@ public:
         OutletFill,
         Text
     };};
-};
-
-struct PhiLookAndFeel  : public juce::LookAndFeel_V4
-{
-    PhiLookAndFeel()
-    {
-        // General
-        setColour(PhiColourIds::General::Background, juce::Colours::darkgrey);
-        setColour(PhiColourIds::General::TopBar,     juce::Colours::darkgrey.darker());
-        setColour(PhiColourIds::General::Highlight,  juce::Colour::greyLevel(0.8f));
-        
-        // Modules (defaults)
-        setColour(PhiColourIds::Module::Highlight,         juce::Colours::cyan.withSaturation(0.5f));
-        setColour(PhiColourIds::Module::Lowlight,          juce::Colour::greyLevel(0.205f));
-        setColour(PhiColourIds::Module::DisabledHighlight, juce::Colours::grey);
-        setColour(PhiColourIds::Module::DisabledText,      juce::Colours::grey);
-        setColour(PhiColourIds::Module::Background,        juce::Colours::darkgrey.darker());
-        setColour(PhiColourIds::Module::Outline,           juce::Colours::grey);
-        setColour(PhiColourIds::Module::SelectedOutline,   juce::Colours::grey.brighter());
-        setColour(PhiColourIds::Module::Name,              juce::Colours::grey);
-        setColour(PhiColourIds::Module::SelectedName,      juce::Colours::grey.brighter());
-        setColour(PhiColourIds::Module::Text,              juce::Colours::grey.brighter());
-        
-        // Ports
-        setColour(PhiColourIds::Port::IntletOutline, juce::Colours::grey);
-        setColour(PhiColourIds::Port::IntletFill,    juce::Colours::darkgrey);
-        setColour(PhiColourIds::Port::OutletOutline, juce::Colours::darkgrey);
-        setColour(PhiColourIds::Port::OutletFill,    juce::Colours::grey);
-        setColour(PhiColourIds::Port::Text,          juce::Colours::grey);
-        
-        // Connections
-        setColour(PhiColourIds::Connection::DefaultFill,     juce::Colours::grey);
-        setColour(PhiColourIds::Connection::SelectedOutline, juce::Colours::grey.brighter());
-        
-        // Other
-        setColour(juce::PopupMenu::highlightedTextColourId,       findColour(PhiColourIds::Module::Text).withMultipliedBrightness(1.3f));
-        setColour(juce::PopupMenu::textColourId,                  findColour(PhiColourIds::Module::Text));
-        setColour(juce::PopupMenu::backgroundColourId,            findColour(PhiColourIds::Module::Background));
-        setColour(juce::PopupMenu::highlightedBackgroundColourId, findColour(PhiColourIds::Module::Lowlight));
-        setColour(juce::TextEditor::focusedOutlineColourId,       juce::Colours::transparentBlack);
-        setColour(juce::TextEditor::highlightedTextColourId,      juce::Colour::greyLevel(0.7f));
-        setColour(juce::ScrollBar::thumbColourId,                 juce::Colour::greyLevel(0.6f));
-        setColour(juce::Slider::textBoxHighlightColourId,         juce::Colour::greyLevel(0.2f));
-        setColour(juce::Slider::textBoxOutlineColourId,           juce::Colours::transparentBlack);
-        setColour(juce::TooltipWindow::backgroundColourId,        juce::Colours::grey);
-        setColour(juce::TooltipWindow::textColourId,              juce::Colours::darkgrey.darker());
-        setColour(juce::ColourSelector::backgroundColourId,       juce::Colours::transparentBlack);
-        setColour(juce::CaretComponent::caretColourId,            juce::Colour::greyLevel(0.8f));
-        setColour(juce::Label::backgroundWhenEditingColourId,     juce::Colour::greyLevel(0.3f));
-    }
-    
-    // This *must* be called after the lookandfeel has been set
-    void hasBeenSet() {
-        //phiLookAndFeel.setDefaultSansSerifTypeface(juce::Typeface::createSystemTypefaceFor(BinaryData::BasierSquareMono_Regular_otf, BinaryData::BasierSquareMono_Regular_otfSize));
-        setDefaultSansSerifTypefaceName ("Helvetica Neue");
-    }
-    
-    void drawCallOutBoxBackground (juce::CallOutBox& box, juce::Graphics& g, const juce::Path&, juce::Image&) override {
-        auto bounds = box.getLocalBounds().reduced(15).toFloat();
-        
-        g.setColour (findColour(PhiColourIds::Module::Background));
-        g.fillRoundedRectangle(bounds, 2.0f);
-        
-        g.setColour (findColour(PhiColourIds::Module::SelectedOutline));
-        g.drawRoundedRectangle(bounds, 2.0f, 2.0f);
-    };
-    
-    int getPopupMenuBorderSize() override { return 1; }
-    
-    void drawLasso(juce::Graphics& g, juce::Component& lasso) override {
-        juce::Path path;
-        path.addRectangle(lasso.getLocalBounds());
-        
-        const float dashes[2] = {5.0f, 5.0f};
-        juce::PathStrokeType(2.0f).createDashedStroke(path, path, dashes, 2);
-        
-        g.setColour(juce::Colours::grey);
-        g.fillPath(path);
-    }
-    
-    void drawMenuBarBackground(juce::Graphics& g, int width, int height, bool isMouseOverBar, juce::MenuBarComponent& menu) override {
-        
-    }
-    
-    void drawMenuBarItem(juce::Graphics& g,
-                         int width, int height,
-                         int itemIndex, const juce::String &itemText,
-                         bool isMouseOverItem,
-                         bool isMenuOpen,
-                         bool isMouseOverBar,
-                         juce::MenuBarComponent& menu) override
-    {
-        // This function uses PopupMenu's colour IDs to maintain a consistent look
-        g.setColour(menu.findColour(isMouseOverItem ? juce::PopupMenu::highlightedTextColourId : juce::PopupMenu::textColourId));
-        g.drawText(itemText, juce::Rectangle<int>{0, 0, width, height}, juce::Justification::centred, 1);
-    }
 };
 
 inline juce::CallOutBox& openColourSelector(juce::Rectangle<int> area, juce::Colour initialColour, juce::Component* comp, juce::ChangeListener* listener)
